requirements.txt
fastapi==0.115.6
pydantic-settings==2.7.0
uvicorn[standard]==0.34.0
httpx==0.28.1
python-telegram-bot==21.10
